Restructure project
At a high level, this commit does the following:

- Renames the project (in the build, repo to follow) to 'hapi-server'.
  A name like 'latis-hapi' is better suited to a project to develop
  HAPI adapters for LaTiS. This is a project to develop a standalone
  HAPI server and HAPI service interface.

- Updates the README to better reflect that this is a standalone HAPI
  server.

- Renames our base package from 'latis.server' to 'lasp.hapi' to
  better reflect that this is a standalone HAPI server and not part of
  LaTiS.

- Splits the server stuff into a 'lasp.hapi.server' package. This is
  the standalone HAPI server.

- Splits the service interface stuff into a 'lasp.hapi.service'
  package. This package could conceivably be made into a separate
  artifact at some point and be used in other servers that require a
  HAPI service interface.

- Splits many things into their own files rather than living where it
  was convenient to put them as they were implemented.

See #17.
