+++ src/main/scala/lasp/hapi/service/CapabilitiesService.scala
@@ -1,11 +1,12 @@
-package latis.server
+package lasp.hapi.service
 
 import cats.effect.Effect
-import io.circe.Encoder
 import io.circe.syntax._
 import org.http4s.HttpService
 import org.http4s.circe._
 import org.http4s.dsl.Http4sDsl
+
+import lasp.hapi.server.HapiServer
 
 /** Implements the `/capabilities` endpoint. */
 class CapabilitiesService[F[_]: Effect] extends Http4sDsl[F] {
@@ -20,27 +21,5 @@ class CapabilitiesService[F[_]: Effect] extends Http4sDsl[F] {
             List("binary", "csv", "json")
           ).asJson
         )
-    }
-}
-
-/**
- * Response from the `/capabilities` endpoint.
- *
- * @param version version of HAPI
- * @param status HAPI status object
- * @param formats list of supported formats
- */
-final case class Capabilities(
-  version: String,
-  status: Status,
-  formats: List[String]
-)
-
-object Capabilities {
-
-  /** JSON encoder */
-  implicit val encoder: Encoder[Capabilities] =
-    Encoder.forProduct3("HAPI", "status", "outputFormats") { x =>
-      (x.version, x.status, x.formats)
     }
 }
