+++ src/main/scala/lasp/hapi/service/Catalog.scala
@@ -1,0 +1,25 @@
+package lasp.hapi.service
+
+import io.circe.Encoder
+
+/**
+ * Representation of a HAPI catalog.
+ *
+ * @param version version of HAPI
+ * @param status HAPI status object
+ * @param catalog list of available datasets
+ */
+final case class Catalog(
+  version: String,
+  status: Status,
+  catalog: List[Dataset]
+)
+
+object Catalog {
+
+  /** JSON encoder */
+  implicit val encoder: Encoder[Catalog] =
+    Encoder.forProduct3("HAPI", "status", "catalog") { x =>
+      (x.version, x.status, x.catalog)
+    }
+}
