+++ src/main/scala/lasp/hapi/service/InfoRequest.scala
@@ -1,0 +1,12 @@
+package lasp.hapi.service
+
+/**
+ * Represents a request to the `info` service.
+ *
+ * @param id HAPI dataset ID
+ * @param parameters parameters to include in response
+ */
+final case class InfoRequest(
+  id: String,
+  parameters: Option[List[String]]
+)
