+++ src/main/scala/lasp/hapi/service/Status.scala
@@ -1,4 +1,4 @@
-package latis.server
+package lasp.hapi.service
 
 import io.circe.Encoder
