+++ src/test/scala/lasp/hapi/service/CapabilitiesServiceSpec.scala
@@ -1,4 +1,4 @@
-package latis.server
+package lasp.hapi.service
 
 import cats.effect.IO
 import io.circe._
@@ -6,6 +6,8 @@ import io.circe.syntax._
 import org.http4s._
 import org.http4s.implicits._
 import org.scalatest.FlatSpec
+
+import lasp.hapi.server.HapiServer
 
 class CapabilitiesServiceSpec extends FlatSpec {
