+++ src/test/scala/lasp/hapi/service/CatalogServiceSpec.scala
@@ -1,4 +1,4 @@
-package latis.server
+package lasp.hapi.service
 
 import org.scalatest.FlatSpec
